code refactoring - in progress
