+++ src/Dto/NoContentResult.cs
@@ -1,0 +1,10 @@
+﻿namespace Morph.Server.Sdk.Dto
+{
+    internal sealed class NoContentResult
+    {
+
+    }
+
+
+
+}
