+++ src/Model/DownloadFileInfo.cs
@@ -6,13 +6,13 @@ using System.Threading.Tasks;
 
 namespace Morph.Server.Sdk.Model
 {
-    public class DownloadFileInfo
-    {
-        /// <summary>
-        /// File name
-        /// </summary>
-        public string FileName { get; set; }
+    //public class DownloadFileInfo
+    //{
+    //    /// <summary>
+    //    /// File name
+    //    /// </summary>
+    //    public string FileName { get; set; }
 
 
-    }
+    //}
 }
