+++ src/Model/InternalModels/OperationType.cs
@@ -1,0 +1,10 @@
+﻿namespace Morph.Server.Sdk.Model.InternalModels
+{
+    internal enum OperationType
+    {
+        ShortOperation = 1,
+        FileTransfer = 2
+    }
+}
+
+
